Refactor to make components reusable
